+++ formatters/array.js
@@ -30,21 +30,23 @@ function arrayJoin (d, separator) {
  * Flatten an array of objects?
  * It ignores nested objects and arrays
  *
- * @example [ [{"id":2, "name":"homer"}, {"id":3, "name":"bart"} ]              ]
- * @example [ [{"id":2, "name":"homer"}, {"id":3, "name":"bart"} ] , " - "      ]
- * @example [ [{"id":2, "name":"homer"}, {"id":3, "name":"bart"} ] , " ; ", "|" ]
- * @example [ [{"id":2, "name":"homer", "obj":{"id":20}, "arr":[12,23] }]       ]
- * @example [ ["homer", "bart", "lisa"]                                         ]
- * @example [ [10, 50]                                                          ]
- * @example [ []                                                                ]
- * @example [ null                                                              ]
- * @example [ {}                                                                ]
- * @example [ 20                                                                ]
- * @example [                                                                   ]
+ * @example [ [{"id":2, "name":"homer"}, {"id":3, "name":"bart"} ]                    ]
+ * @example [ [{"id":2, "name":"homer"}, {"id":3, "name":"bart"} ] , " - "            ]
+ * @example [ [{"id":2, "name":"homer"}, {"id":3, "name":"bart"} ] , " ; ", "|"       ]
+ * @example [ [{"id":2, "name":"homer"}, {"id":3, "name":"bart"} ] , " ; ", "|", "id" ]
+ * @example [ [{"id":2, "name":"homer", "obj":{"id":20}, "arr":[12,23] }]             ]
+ * @example [ ["homer", "bart", "lisa"]                                               ]
+ * @example [ [10, 50]                                                                ]
+ * @example [ []                                                                      ]
+ * @example [ null                                                                    ]
+ * @example [ {}                                                                      ]
+ * @example [ 20                                                                      ]
+ * @example [                                                                         ]
  *
  * @param  {Array} d                     array passed by carbone
  * @param  {String} objSeparator         [optional] object separator (`, ` by default)
  * @param  {String} attributeSeparator   [optional] attribute separator (`:` by default)
+ * @param  {String [, ...]} attributes   [optional] list of object's attributes to print
  * @return {String}                      the computed result, or `d` if `d` is not an array
  */
 function arrayMap (d, objSeparator, attributeSeparator) {